Remove erroneous treasury link
I have no idea how I got so confused about this, but I added a Google Docs form URL to the treasury page when it needed to be in the hotel setup page. It's already in the hotel setup page, so we just need to remove the bad treasury link.