Pin Super 2019 to appropriate branches
A deploy was run against the old Super instance, causing it to break due to the extreme changes in the code. I've made branches from the latest commits in January for both ubersystem and magprime, and this will pin Super 2019 to those branches so we can get that server back to the proper state.