Make staff and contractor badge ranges mergeable
We want contractor badges to be a separate range from staff badges _for our badge export_ so that we can easily put "Contractor" stickers on those badges. HOWEVER, when it comes to assigning blank badges, we want to use one range of blank badges for both Staff and Contractors.

Our plan was to use separate ranges for the export and then merge them afterwards. The catch here is that the system doesn't support two sets of ranges -- because the guest badge range was splitting staff and contractor ranges, we wouldn't have been able to merge the badge types' ranges properly. So to fix that we're rejiggering the ranges before the export. After the export, both staff and contractor badges will use the range 25, 2999.